moved get_adjacency_types to lifter
